+++ tests/AddDynamicsRole.ps1
@@ -1,0 +1,35 @@
+$dbHostName = $env:VMDEVOPSSTARTER_DBHOST;
+if ( !$dbHostName ) { $dbHostName = $env:COMPUTERNAME }
+$securedPassword = ConvertTo-SecureString "c0mp1Expa~~" -AsPlainText -Force
+$CRMInstallAccountCredential = New-Object System.Management.Automation.PSCredential( "contoso\_crmadmin", $securedPassword );
+
+$domainName = (Get-WmiObject Win32_ComputerSystem).Domain;
+
+try {
+    Write-Host "Invoking command on $env:COMPUTERNAME.$domainName with dbHostName=$dbHostName parameter";
+    Invoke-Command "$env:COMPUTERNAME.$domainName" -Credential $CRMInstallAccountCredential -Authentication CredSSP {
+        param( $dbHostName )
+        Write-Host "Invoked, starting execution";
+        Import-Module c:/test-projects/Dynamics365Configuration/src/Dynamics365Configuration/Dynamics365Configuration.psd1;
+        $securedPassword = ConvertTo-SecureString "c0mp1Expa~~" -AsPlainText -Force
+        $CRMServiceAccountCredential = New-Object System.Management.Automation.PSCredential( "contoso\_crmsrv", $securedPassword );
+        $MonitoringServiceAccountCredential = New-Object System.Management.Automation.PSCredential( "contoso\_crmmon", $securedPassword );
+        Write-Host "Joining Dynamics config db on $dbHostName";
+        Install-Dynamics365Server `
+            -MediaDir C:\Install\Dynamics\Dynamics365Server90RTMEnu `
+            -LicenseKey KKNV2-4YYK8-D8HWD-GDRMW-29YTW `
+            -ServerRoles FrontEnd `
+            -SqlServer $dbHostName\SQLInstance01 `
+            -CrmServiceAccount $CRMServiceAccountCredential `
+            -MonitoringServiceAccount $MonitoringServiceAccountCredential `
+            -LogFilePath c:\tmp\Dynamics365ServerInstallLog.txt `
+            -LogFilePullIntervalInSeconds 15 `
+            -LogFilePullToOutput
+    } -ArgumentList $dbHostName;
+} catch {
+    Write-Host "Failed in invoking of Install-Dynamics365Server";
+    Write-Host $_.Exception.Message -ForegroundColor Red;
+    Exit 1;
+}
+
+Exit 0;